requirements.txt
frame-ble==1.1.1
frame-msg==5.2.0
pillow==11.1.0